requirements.txt
# Python包依赖
pyyaml>=6.0
setuptools>=45.0.0
wheel>=0.34.0
psutil>=5.8.0

# 生物信息学软件依赖
# 以下软件可以通过conda/mamba安装：
# mamba install -c bioconda:
# - gatk4
# - bwa
# - samtools
# - picard
# - vcftools
# - bgzip
# - tabix
# - fastp
# - qualimap
# - multiqc

# 需要手动安装的软件：
# - bwa-mem2 (https://github.com/bwa-mem2/bwa-mem2) 